Only backup shortcuts if the user has shortcuts

+++ UWPHook/GamesWindow.xaml.cs
@@ -554,6 +554,12 @@ namespace UWPHook
             string sourceFilePath = Path.Combine(userPath, "config", "shortcuts.vdf");
             string destinationFileName = Path.Combine(backupFolder, $"{user_id}_{DateTime.Now.ToString("yyyyMMddHHmmss")}_shortcuts.vdf");
 
+            if (!File.Exists(sourceFilePath))
+            {
+               Log.Information("No VDF file found for user: " + sourceFilePath);
+               return;
+            }
+
             File.Copy(sourceFilePath, destinationFileName);
             Log.Debug("Backup created: " + destinationFileName);
         }
